% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-meth.R
\name{extract_meth_at_context}
\alias{extract_meth_at_context}
\title{Extract methylation at an intervals set}
\usage{
extract_meth_at_context(intervals, iterator = intervals, min_cov = NULL)
}
\arguments{
\item{intervals}{intervals to extract methylation at}

\item{iterator}{iterator of gextract_meth call}

\item{min_cov}{minimal coverage}
}
\description{
Extract methylation at an intervals set
}
