% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAD-methylation.R
\name{get_cpg_meth_tads}
\alias{get_cpg_meth_tads}
\title{Get 'background' methylation in TADs}
\usage{
get_cpg_meth_tads(
  tad_intervs,
  meth_dips,
  k4me3_peaks,
  k27me3_peaks,
  atac_peaks
)
}
\arguments{
\item{tad_intervs}{intervals of the TADs}

\item{meth_dips, k4me3_peaks, k27me3_peaks, atac_peaks}{intervals of functional annotations (see notebook)}
}
\description{
Get 'background' methylation in TADs
}
