% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth-dips.R
\name{meth_dip_screen}
\alias{meth_dip_screen}
\title{Screen for "dips" in methylation}
\usage{
meth_dip_screen(
  track,
  intervals = gintervals.all(),
  iterator = "intervs.global.seq_CG",
  dip_size = 300,
  flank_size = 500,
  margin = 50,
  min_diff = 0.5,
  min_cov = 10,
  min_nhood_cov = min_cov,
  canonize = FALSE,
  ret_values = FALSE
)
}
\arguments{
\item{track}{methylation track prefix (without the ".cov" or ".meth" suffix)}

\item{intervals}{intervals set}

\item{iterator}{iterator. Default - CpGs}

\item{dip_size}{size of the dip region}

\item{flank_size}{size of flank regions}

\item{margin}{margin between the dip region and the flanking regions}

\item{min_diff}{minimal average methylation difference between dip regions and flanking regions. A "dip" would be called only if a difference exists between the center and both right and left flanks.}

\item{min_cov}{minimal coverage for center}

\item{min_nhood_cov}{minimal coverage for flanks}

\item{canonize}{merge overlapping regions in the resulting intervals}

\item{ret_values}{if FALSE - returns only the "dip" intervals. if TRUE - returns also the "cov" and "meth" values for the center and flanking regions with intervals shifted by dip_size/2.}
}
\description{
Screen for "dips" in methylation
}
