% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_smooth_cpg_scatter}
\alias{plot_smooth_cpg_scatter}
\title{Plot a smooth scatter of methylation data}
\usage{
plot_smooth_cpg_scatter(
  df,
  col1,
  col2,
  min_cov,
  xlab = col1,
  ylab = col2,
  colors = c("white", "lightblue", "blue", "darkblue", "yellow", "gold", "orange",
    "red", "darkred")
)
}
\arguments{
\item{df}{Dataframe with methylation data}

\item{col1}{Column name for the first dataset}

\item{col2}{Column name for the second dataset}

\item{min_cov}{Minmial coverage for a site to be included in the plot}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{colors}{Vector of colors for the scatter}
}
\description{
Plot a smooth scatter of methylation data
}
