% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_smooth_legend}
\alias{plot_smooth_legend}
\title{Plot a legend for smooth scatter}
\usage{
plot_smooth_legend(colors, n_colors = 1000)
}
\arguments{
\item{colors}{Vector of colors}

\item{n_colors}{Number of colors in the legend}
}
\description{
Plot a legend for smooth scatter
}
\examples{
plot_smooth_legend(c("white", "lightblue", "blue", "darkblue", "yellow", "gold", "orange", "red", "darkred"), 1000)

}
