% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tet.gastru-package.R
\docType{package}
\name{tet.gastru-package}
\alias{tet.gastru}
\alias{tet.gastru-package}
\title{tet.gastru: The intrinsic and extrinsic effects of Tet proteins during gastrulation}
\description{
Functions for supporting analysis from the paper: "The intrinsic and extrinsic effects of Tet proteins during gastrulation", Cell 2022
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
